+++ packages/angular-eslint/src/configs/template-all.ts
@@ -43,6 +43,7 @@ export default (
       '@angular-eslint/template/prefer-at-else': 'error',
       '@angular-eslint/template/prefer-at-empty': 'error',
       '@angular-eslint/template/prefer-built-in-pipes': 'error',
+      '@angular-eslint/template/prefer-class-binding': 'error',
       '@angular-eslint/template/prefer-contextual-for-variables': 'error',
       '@angular-eslint/template/prefer-control-flow': 'error',
       '@angular-eslint/template/prefer-ngsrc': 'error',