+++ packages/eslint-plugin/src/index.ts
@@ -142,6 +142,12 @@ import noExperimental, {
 import noDeveloperPreview, {
   RULE_NAME as noDeveloperPreviewRuleName,
 } from './rules/no-developer-preview';
+import noImplicitTakeUntilDestroyed, {
+  RULE_NAME as noImplicitTakeUntilDestroyedRuleName,
+} from './rules/no-implicit-take-until-destroyed';
+import computedMustReturn, {
+  RULE_NAME as computedMustReturnRuleName,
+} from './rules/computed-must-return';
 
 export = {
   configs: {
@@ -152,6 +158,7 @@ export = {
     [componentClassSuffixRuleName]: componentClassSuffix,
     [componentMaxInlineDeclarationsRuleName]: componentMaxInlineDeclarations,
     [componentSelectorRuleName]: componentSelector,
+    [computedMustReturnRuleName]: computedMustReturn,
     [consistentComponentStylesRuleName]: consistentComponentStyles,
     [contextualDecoratorRuleName]: contextualDecorator,
     [contextualLifecycleRuleName]: contextualLifecycle,
@@ -160,30 +167,33 @@ export = {
     [noAsyncLifecycleMethodRuleName]: noAsyncLifecycleMethod,
     [noAttributeDecoratorRuleName]: noAttributeDecorator,
     [noConflictingLifecycleRuleName]: noConflictingLifecycle,
+    [noDeveloperPreviewRuleName]: noDeveloperPreview,
     [noDuplicatesInMetadataArraysRuleName]: noDuplicatesInMetadataArrays,
     [noEmptyLifecycleMethodRuleName]: noEmptyLifecycleMethod,
+    [noExperimentalRuleName]: noExperimental,
     [noForwardRefRuleName]: noForwardRef,
+    [noImplicitTakeUntilDestroyedRuleName]: noImplicitTakeUntilDestroyed,
     [noInputPrefixRuleName]: noInputPrefix,
     [noInputRenameRuleName]: noInputRename,
     [noInputsMetadataPropertyRuleName]: noInputsMetadataProperty,
     [noLifecycleCallRuleName]: noLifecycleCall,
-    [noUncalledSignalsRuleName]: noUncalledSignals,
     [noOutputNativeRuleName]: noOutputNative,
     [noOutputOnPrefixRuleName]: noOutputOnPrefix,
     [noOutputRenameRuleName]: noOutputRename,
     [noOutputsMetadataPropertyRuleName]: noOutputsMetadataProperty,
     [noPipeImpureRuleName]: noPipeImpure,
     [noQueriesMetadataPropertyRuleName]: noQueriesMetadataProperty,
+    [noUncalledSignalsRuleName]: noUncalledSignals,
     [pipePrefixRuleName]: pipePrefix,
     [preferHostMetadataPropertyRuleName]: preferHostMetadataProperty,
+    [preferInjectRuleName]: preferInject,
     [preferOnPushComponentChangeDetectionRuleName]:
       preferOnPushComponentChangeDetection,
-    [preferSignalsRuleName]: preferSignals,
-    [preferStandaloneRuleName]: preferStandalone,
-    [preferInjectRuleName]: preferInject,
     [preferOutputEmitterRefRuleName]: preferOutputEmitterRef,
     [preferOutputReadonlyRuleName]: preferOutputReadonly,
     [preferSignalModelRuleName]: preferSignalModel,
+    [preferSignalsRuleName]: preferSignals,
+    [preferStandaloneRuleName]: preferStandalone,
     [relativeUrlPrefixRuleName]: relativeUrlPrefix,
     [requireLifecycleOnPrototypeRuleName]: requireLifecycleOnPrototype,
     [requireLocalizeMetadataRuleName]: requireLocalizeMetadata,
@@ -195,7 +205,5 @@ export = {
     [useInjectableProvidedInRuleName]: useInjectableProvidedIn,
     [useLifecycleInterfaceRuleName]: useLifecycleInterface,
     [usePipeTransformInterfaceRuleName]: usePipeTransformInterface,
-    [noExperimentalRuleName]: noExperimental,
-    [noDeveloperPreviewRuleName]: noDeveloperPreview,
   },
 };