+++ packages/eslint-plugin/tests/rules/no-implicit-take-until-destroyed/spec.ts
@@ -1,0 +1,12 @@
+import { RuleTester } from '@angular-eslint/test-utils';
+import rule, {
+  RULE_NAME,
+} from '../../../src/rules/no-implicit-take-until-destroyed';
+import { invalid, valid } from './cases';
+
+const ruleTester = new RuleTester();
+
+ruleTester.run(RULE_NAME, rule, {
+  valid,
+  invalid,
+});